Change actor position to follow UI Area Size instead of Screen Size
RPG Maker MZ comes with a new feature, where the UI Area size can be different from the Screen size. This is very helpful if we are making a game with multi aspect ratio.
SRPG should adopt this feature so that the battle view remains consistent regardless of the aspect ratio displayed.