Disable  cursorFollowMouse during 'actor_phase' and 'normal' sub battle phase
Hi, Mr. Takumi Ariake.
Please review this and make adjustments if necessary.

By default, if the parameters cursorFollowMouse = 'true', and isWheelPrevNext = 'true', the mouse wheel function will be disrupted because the cursor (player) will always follow the position of the mouse pointer. With this patch, as long as the condition _isBattlePhase = 'actor_phase' and the _isSubBattlePhase = 'normal', the function cursorFollowMouse = 'true' will remain 'false'.

My patch hasn't solved the problem where at the beginning of _isBattlePhase = 'actor_phase' and _isSubBattlePhase = 'actor_target' there was a brief moment where the cursor was switching between following the mouse pointer position or following the actor position.

In terms of user experience, this does not look good. But it seems that in the section you commented on code lines 434 - 470 there is a function to solve this problem.

This plugin also still has a bug, when using touch after using a mouse, the cursorFollowMouse function remains 'true' so that the cursor will jump repeatedly between the position of the selected actor and the current mouse pointer position (tested with a laptop that has a touchscreen feature), while a similar bug is not found if using a keyboard.